pyproject.toml
[tool.poetry]
name = "drf-kurs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "drf_kurs"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.8"
pillow = "^10.0.1"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.0"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
redis = "^5.0.1"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
